Add support for 'format' plugin option.
This commit adds support for a user-supplied set of fields to get
`git log` information for instead of a fixed set.  The primary
motivation for this change was to support getting "commit date" instead
of author date, but it seemed better to create a more flexible solution
rather than focusing on a single new field.

For more details see the updated README.md file.